+++ kitty_tests/mouse.py
@@ -79,8 +79,9 @@ class TestMouse(BaseTest):
         def move(x=0, y=0, button=-1, q=None):
             ev(x=x, y=y, button=button)
             if q is not None:
-                s = sel()
-                self.ae(s, q, f'{s!r} != {q!r} after movement to x={x} y={y}')
+                sl = sel()
+                from kitty.window import as_text
+                self.ae(sl, q, f'{sl!r} != {q!r} after movement to x={x} y={y}. Screen contents: {as_text(s)!r}')
 
         def multi_click(x=0, y=0, count=2):
             clear_click_queue = True
@@ -277,3 +278,12 @@ class TestMouse(BaseTest):
         move(x=3.6, y=2, q='abcd')
         press(x=3, y=0, button=GLFW_MOUSE_BUTTON_RIGHT)
         self.ae(sel(), '4567890abcd')
+
+        # blank line select
+        s.reset()
+        s.draw('abcde')
+        s.linefeed(), s.carriage_return()
+        s.linefeed(), s.carriage_return()
+        s.draw('12345')
+        press(x=0, y=0)
+        move(x=2, y=2, q='abcde\n\n12')