+++ docs/changelog.rst
@@ -101,6 +101,8 @@ Detailed list of changes
 
 - A new option :opt:`clear_selection_on_clipboard_loss` to clear selections when they no longer reflect the contents of the clipboard
 
+- Fix a regression in the previous release that caused empty lines to be skipped when copying text from a selection (:iss:`8435`)
+
 - Fix flickering of hyperlink underline when client program continuously
   redraws on mouse movement (:iss:`8414`)